projects/labelstudio/pyproject.toml
# Purpose: Declare the Label Studio subproject package metadata and install surface.
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "birdsys-labelstudio"
version = "0.1.0"
description = "BirdSys Label Studio workflows"
requires-python = ">=3.11,<3.12"
dependencies = [
  "matplotlib>=3.9,<4",
  "pandas>=2.2,<3",
  "pillow>=10,<12",
  "pyarrow>=18,<20",
]

[project.scripts]
birdsys-labelstudio-extract-annotations = "birdsys.labelstudio.extract_annotations:main"
birdsys-labelstudio-build-localfiles-batch = "birdsys.labelstudio.build_localfiles_batch:main"
birdsys-labelstudio-sync-species-data = "birdsys.labelstudio.sync_species_data:main"
birdsys-labelstudio-backfill-task-metadata = "birdsys.labelstudio.backfill_task_metadata:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["birdsys.labelstudio*"]
namespaces = true
